BJS2-78371 parallel save hash

+++ src/main/java/faang/school/urlshortenerservice/repository/HashRepository.java
@@ -31,5 +31,5 @@ public interface HashRepository extends JpaRepository<Hash, String> {
                            LIMIT :count)
             RETURNING *
             """)
-    List<Hash> findAndDeleteLimit(@Param("count") int count);
+    List<Hash> findAndDeleteLimit(@Param("count") long count);
 }
